+++ packages/protos-runtime-py/src/protos_runtime/models/__init__.py
@@ -1,0 +1,6 @@
+"""Pydantic contract models extracted from semcod/protos generated/.
+
+Public submodules:
+    - search_v1
+    - identification_v1
+"""